+++ vcd/config_test.go
@@ -43,8 +43,13 @@ type TestConfig struct {
 		MaxRetryTimeout          int    `json:"maxRetryTimeout"`
 	} `json:"provider"`
 	VCD struct {
-		Org     string `json:"org"`
-		Vdc     string `json:"vdc"`
+		Org         string `json:"org"`
+		Vdc         string `json:"vdc"`
+		ProviderVdc struct {
+			Name           string `json:"name"`
+			NetworkPool    string `json:"networkPool"`
+			StorageProfile string `json:"storageProfile"`
+		} `json:"providerVdc"`
 		Catalog struct {
 			Name        string `json:"name,omitempty"`
 			CatalogItem string `json:"catalogItem,omitempty"`